% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbaModel.R
\docType{package}
\name{lbaModel}
\alias{lbaModel-package}
\alias{lbaModel}
\title{The Linear Ballistic Accumulaion}
\description{
Provides density, distribution and random generation functions for the
Linear Ballistic Accumulator (LBA) model, a widely used choice response
time model in cognitive psychology. The package supports model
specifications, parameter estimation, and likelihood computation,
facilitating simulation and statistical inference for LBA-based experiments.
For background on the LBA model, see Brown and Heathcote (2008)
<doi:10.1016/j.cogpsych.2007.12.002>.
}
\author{
Yi-Shin Lin <yishinlin001@gmail.com> \cr
}
\keyword{internal}
\keyword{package}
